Actualizacion de base de datos con tablas gig y genre
