+++ Code/GitHub/GitHub/GitHub.csproj
@@ -202,11 +202,20 @@
     <Compile Include="Migrations\201707250611193_InitialMigration.Designer.cs">
       <DependentUpon>201707250611193_InitialMigration.cs</DependentUpon>
     </Compile>
+    <Compile Include="Migrations\201707251603235_CreateGigTable.cs" />
+    <Compile Include="Migrations\201707251603235_CreateGigTable.Designer.cs">
+      <DependentUpon>201707251603235_CreateGigTable.cs</DependentUpon>
+    </Compile>
+    <Compile Include="Migrations\201707251608584_OverrideConventionsForGigsAndGenres.cs" />
+    <Compile Include="Migrations\201707251608584_OverrideConventionsForGigsAndGenres.Designer.cs">
+      <DependentUpon>201707251608584_OverrideConventionsForGigsAndGenres.cs</DependentUpon>
+    </Compile>
     <Compile Include="Migrations\Configuration.cs" />
     <Compile Include="Models\AccountViewModels.cs" />
+    <Compile Include="Models\ApplicationDbContext.cs" />
     <Compile Include="Models\Genre.cs" />
     <Compile Include="Models\Gig.cs" />
-    <Compile Include="Models\IdentityModels.cs" />
+    <Compile Include="Models\ApplicationUser.cs" />
     <Compile Include="Models\ManageViewModels.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Startup.cs" />
@@ -298,6 +307,12 @@
   <ItemGroup>
     <EmbeddedResource Include="Migrations\201707250611193_InitialMigration.resx">
       <DependentUpon>201707250611193_InitialMigration.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="Migrations\201707251603235_CreateGigTable.resx">
+      <DependentUpon>201707251603235_CreateGigTable.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="Migrations\201707251608584_OverrideConventionsForGigsAndGenres.resx">
+      <DependentUpon>201707251608584_OverrideConventionsForGigsAndGenres.cs</DependentUpon>
     </EmbeddedResource>
   </ItemGroup>
   <PropertyGroup>
