+++ Code/GitHub/GitHub/Models/ApplicationDbContext.cs
@@ -1,0 +1,22 @@
+using System.Data.Entity;
+using Microsoft.AspNet.Identity.EntityFramework;
+
+namespace GitHub.Models
+{
+    public class ApplicationDbContext : IdentityDbContext<ApplicationUser>
+    {
+        public DbSet<Gig> Gigs { get; set; }
+
+        public DbSet<Genre> Genres { get; set; }
+
+        public ApplicationDbContext()
+            : base("DefaultConnection", throwIfV1Schema: false)
+        {
+        }
+
+        public static ApplicationDbContext Create()
+        {
+            return new ApplicationDbContext();
+        }
+    }
+}
