+++ Code/GitHub/GitHub/Models/Genre.cs
@@ -1,9 +1,13 @@
+using System.ComponentModel.DataAnnotations;
+
 namespace GitHub.Models
 {
     public class Genre
     {
         public byte Id { get; set; }
 
+        [Required]
+        [StringLength(255)]
         public string Name { get; set; }
     }
 }
