+++ Code/GitHub/GitHub/Models/Gig.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Web;
 
@@ -9,12 +10,17 @@ namespace GitHub.Models
     {
         public int Id { get; set; }
 
+        [Required]
         public ApplicationUser Artist { get; set; }
+
 
         public DateTime DateTime { get; set; }
 
+        [Required]
+        [StringLength(255)]
         public string Venue { get; set; }
 
+        [Required]
         public Genre Genre { get; set; }
     }
 }
